CI: Overhaul build matrix for Ubuntu runners.
The complexity of the build matrix for the Ubuntu runners grew to a level
which made it very difficult to understand which set of runners will be
started with which configuration.

Overhaul the build matrix settings by moving most of its complexity to the
build steps.
That should make it easier to potentially adjust the settings for one of
the runners in the matrix in the future.